backend/requirements.txt
fastapi
uvicorn
python-multipart
email-validator
spacy
aiofiles
python-dotenv
